FIX : Align nodes patch by Chris Baird (closes #2296)
git-svn-id: http://svn.openstreetmap.org/applications/editors/merkaartor@17738 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ src/Maps/Coord.h
@@ -80,7 +80,7 @@ class Coord
 
 		double length() const
 		{
-			return sqrt((double)(Lat*Lat+Lon*Lon));
+			return sqrt((double)((long)Lat*(long)Lat+(long)Lon*(long)Lon));
 		}
 
 		double distanceFrom(const Coord& other) const;
